Add current decompile to show progress

+++ libntr/include/ntr/base.h
@@ -8,7 +8,7 @@
 
 #include "base.asm.h"
 
-typedef uint32_t u8;
+typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
 typedef uint64_t u64;
